Rename apply route personal-information to contact-information
