+++ frontend/app/routes/$lang/_public/apply/$id/adult-child/applicant-information.tsx
@@ -145,7 +145,7 @@ export async function action({ context: { session }, params, request }: ActionFu
     return redirect(getPathById('$lang/_public/apply/$id/adult-child/partner-information', params));
   }
 
-  return redirect(getPathById('$lang/_public/apply/$id/adult-child/personal-information', params));
+  return redirect(getPathById('$lang/_public/apply/$id/adult-child/contact-information', params));
 }
 
 export default function ApplyFlowApplicationInformation() {
