+++ frontend/app/routes/$lang/_public/apply/$id/adult-child/review-adult-information.tsx
@@ -325,7 +325,7 @@ export default function ReviewInformation() {
               <DescriptionListItem term={t('apply-adult-child:review-adult-information.phone-title')}>
                 {userInfo.phoneNumber}
                 <p className="mt-4">
-                  <InlineLink id="change-phone-number" routeId="$lang/_public/apply/$id/adult-child/personal-information" params={params}>
+                  <InlineLink id="change-phone-number" routeId="$lang/_public/apply/$id/adult-child/contact-information" params={params}>
                     {t('apply-adult-child:review-adult-information.phone-change')}
                   </InlineLink>
                 </p>
@@ -333,7 +333,7 @@ export default function ReviewInformation() {
               <DescriptionListItem term={t('apply-adult-child:review-adult-information.alt-phone-title')}>
                 {userInfo.altPhoneNumber}
                 <p className="mt-4">
-                  <InlineLink id="change-alternate-phone-number" routeId="$lang/_public/apply/$id/adult-child/personal-information" params={params}>
+                  <InlineLink id="change-alternate-phone-number" routeId="$lang/_public/apply/$id/adult-child/contact-information" params={params}>
                     {t('apply-adult-child:review-adult-information.alt-phone-change')}
                   </InlineLink>
                 </p>
@@ -341,7 +341,7 @@ export default function ReviewInformation() {
               <DescriptionListItem term={t('apply-adult-child:review-adult-information.email')}>
                 {userInfo.contactInformationEmail}
                 <p className="mt-4">
-                  <InlineLink id="change-email" routeId="$lang/_public/apply/$id/adult-child/personal-information" params={params}>
+                  <InlineLink id="change-email" routeId="$lang/_public/apply/$id/adult-child/contact-information" params={params}>
                     {t('apply-adult-child:review-adult-information.email-change')}
                   </InlineLink>
                 </p>
@@ -357,7 +357,7 @@ export default function ReviewInformation() {
                   altFormat={true}
                 />
                 <p className="mt-4">
-                  <InlineLink id="change-mailing-address" routeId="$lang/_public/apply/$id/adult-child/personal-information" params={params}>
+                  <InlineLink id="change-mailing-address" routeId="$lang/_public/apply/$id/adult-child/contact-information" params={params}>
                     {t('apply-adult-child:review-adult-information.mailing-change')}
                   </InlineLink>
                 </p>
@@ -373,7 +373,7 @@ export default function ReviewInformation() {
                   altFormat={true}
                 />
                 <p className="mt-4">
-                  <InlineLink id="change-home-address" routeId="$lang/_public/apply/$id/adult-child/personal-information" params={params}>
+                  <InlineLink id="change-home-address" routeId="$lang/_public/apply/$id/adult-child/contact-information" params={params}>
                     {t('apply-adult-child:review-adult-information.home-change')}
                   </InlineLink>
                 </p>
