+++ frontend/app/routes/$lang/_public/apply/$id/adult/communication-preference.tsx
@@ -275,7 +275,7 @@ export default function ApplyFlowCommunicationPreferencePage() {
                 {t('apply-adult:communication-preference.continue')}
                 <FontAwesomeIcon icon={isSubmitting ? faSpinner : faChevronRight} className={cn('ms-3 block size-4', isSubmitting && 'animate-spin')} />
               </Button>
-              <ButtonLink id="back-button" routeId="$lang/_public/apply/$id/adult/personal-information" params={params} disabled={isSubmitting} data-gc-analytics-customclick="ESDC-EDSC:CDCP Online Application Form-Adult:Back - Communication click">
+              <ButtonLink id="back-button" routeId="$lang/_public/apply/$id/adult/contact-information" params={params} disabled={isSubmitting} data-gc-analytics-customclick="ESDC-EDSC:CDCP Online Application Form-Adult:Back - Communication click">
                 <FontAwesomeIcon icon={faChevronLeft} className="me-3 block size-4" />
                 {t('apply-adult:communication-preference.back')}
               </ButtonLink>
