+++ frontend/app/routes/$lang/_public/apply/$id/adult/contact-information.tsx
@@ -16,9 +16,10 @@ import { ErrorSummary, createErrorSummaryItems, hasErrors, scrollAndFocusToError
 import { InputCheckbox } from '~/components/input-checkbox';
 import { InputField } from '~/components/input-field';
 import { InputOptionProps } from '~/components/input-option';
+import { InputPhoneField } from '~/components/input-phone-field';
 import { InputSelect } from '~/components/input-select';
 import { Progress } from '~/components/progress';
-import { loadApplyAdultChildState } from '~/route-helpers/apply-adult-child-route-helpers.server';
+import { loadApplyAdultState } from '~/route-helpers/apply-adult-route-helpers.server';
 import { ContactInformationState, saveApplyState } from '~/route-helpers/apply-route-helpers.server';
 import { getLookupService } from '~/services/lookup-service.server';
 import * as adobeAnalytics from '~/utils/adobe-analytics.client';
@@ -33,9 +34,9 @@ import { getTitleMetaTags } from '~/utils/seo-utils';
 import { cn } from '~/utils/tw-utils';
 
 export const handle = {
-  i18nNamespaces: getTypedI18nNamespaces('apply-adult-child', 'apply', 'gcweb'),
-  pageIdentifier: pageIds.public.apply.adultChild.contactInformation,
-  pageTitleI18nKey: 'apply-adult-child:contact-information.page-title',
+  i18nNamespaces: getTypedI18nNamespaces('apply-adult', 'apply', 'gcweb'),
+  pageIdentifier: pageIds.public.apply.adult.contactInformation,
+  pageTitleI18nKey: 'apply-adult:contact-information.page-title',
 } as const satisfies RouteHandleData;
 
 export const meta: MetaFunction<typeof loader> = mergeMeta(({ data }) => {
@@ -44,7 +45,7 @@ export const meta: MetaFunction<typeof loader> = mergeMeta(({ data }) => {
 
 export async function loader({ context: { session }, params, request }: LoaderFunctionArgs) {
   const lookupService = getLookupService();
-  const state = loadApplyAdultChildState({ params, request, session });
+  const state = loadApplyAdultState({ params, request, session });
   const t = await getFixedT(request, handle.i18nNamespaces);
   const { CANADA_COUNTRY_ID, USA_COUNTRY_ID, MARITAL_STATUS_CODE_COMMONLAW, MARITAL_STATUS_CODE_MARRIED } = getEnv();
 
@@ -52,7 +53,7 @@ export async function loader({ context: { session }, params, request }: LoaderFu
   const regionList = lookupService.getAllRegions();
 
   const csrfToken = String(session.get('csrfToken'));
-  const meta = { title: t('gcweb:meta.title.template', { title: t('apply-adult-child:contact-information.page-title') }) };
+  const meta = { title: t('gcweb:meta.title.template', { title: t('apply-adult:contact-information.page-title') }) };
 
   return json({
     id: state.id,
@@ -71,9 +72,9 @@ export async function loader({ context: { session }, params, request }: LoaderFu
 }
 
 export async function action({ context: { session }, params, request }: ActionFunctionArgs) {
-  const log = getLogger('apply/adult-child/personal-information');
+  const log = getLogger('apply/contact-information');
 
-  const state = loadApplyAdultChildState({ params, request, session });
+  const state = loadApplyAdultState({ params, request, session });
   const t = await getFixedT(request, handle.i18nNamespaces);
   const { CANADA_COUNTRY_ID, USA_COUNTRY_ID, COMMUNICATION_METHOD_EMAIL_ID } = getEnv();
 
@@ -83,21 +84,21 @@ export async function action({ context: { session }, params, request }: ActionFu
         .string()
         .trim()
         .max(100)
-        .refine((val) => !val || isValidPhoneNumber(val, 'CA'), t('apply-adult-child:contact-information.error-message.phone-number-valid'))
+        .refine((val) => !val || isValidPhoneNumber(val, 'CA'), t('apply-adult:contact-information.error-message.phone-number-valid'))
         .optional(),
       phoneNumberAlt: z
         .string()
         .trim()
         .max(100)
-        .refine((val) => !val || isValidPhoneNumber(val, 'CA'), t('apply-adult-child:contact-information.error-message.phone-number-alt-valid'))
+        .refine((val) => !val || isValidPhoneNumber(val, 'CA'), t('apply-adult:contact-information.error-message.phone-number-alt-valid'))
         .optional(),
       email: z.string().trim().max(64).optional(),
       confirmEmail: z.string().trim().max(64).optional(),
-      mailingAddress: z.string().trim().min(1, t('apply-adult-child:contact-information.error-message.mailing-address.address-required')).max(30),
+      mailingAddress: z.string().trim().min(1, t('apply-adult:contact-information.error-message.mailing-address.address-required')).max(30),
       mailingApartment: z.string().trim().max(30).optional(),
-      mailingCountry: z.string().trim().min(1, t('apply-adult-child:contact-information.error-message.mailing-address.country-required')),
-      mailingProvince: z.string().trim().min(1, t('apply-adult-child:contact-information.error-message.mailing-address.province-required')).optional(),
-      mailingCity: z.string().trim().min(1, t('apply-adult-child:contact-information.error-message.mailing-address.city-required')).max(100),
+      mailingCountry: z.string().trim().min(1, t('apply-adult:contact-information.error-message.mailing-address.country-required')),
+      mailingProvince: z.string().trim().min(1, t('apply-adult:contact-information.error-message.mailing-address.province-required')).optional(),
+      mailingCity: z.string().trim().min(1, t('apply-adult:contact-information.error-message.mailing-address.city-required')).max(100),
       mailingPostalCode: z.string().trim().max(100).optional(),
       copyMailingAddress: z.boolean(),
       homeAddress: z.string().trim().max(30).optional(),
@@ -110,55 +111,55 @@ export async function action({ context: { session }, params, request }: ActionFu
     .superRefine((val, ctx) => {
       if (val.email ?? val.confirmEmail) {
         if (typeof val.email !== 'string' || validator.isEmpty(val.email)) {
-          ctx.addIssue({ code: z.ZodIssueCode.custom, message: t('apply-adult-child:contact-information.error-message.email-required'), path: ['email'] });
+          ctx.addIssue({ code: z.ZodIssueCode.custom, message: t('apply-adult:contact-information.error-message.email-required'), path: ['email'] });
         } else if (!validator.isEmail(val.email)) {
-          ctx.addIssue({ code: z.ZodIssueCode.custom, message: t('apply-adult-child:contact-information.error-message.email-valid'), path: ['email'] });
+          ctx.addIssue({ code: z.ZodIssueCode.custom, message: t('apply-adult:contact-information.error-message.email-valid'), path: ['email'] });
         }
 
         if (typeof val.confirmEmail !== 'string' || validator.isEmpty(val.confirmEmail)) {
-          ctx.addIssue({ code: z.ZodIssueCode.custom, message: t('apply-adult-child:contact-information.error-message.confirm-email-required'), path: ['confirmEmail'] });
+          ctx.addIssue({ code: z.ZodIssueCode.custom, message: t('apply-adult:contact-information.error-message.confirm-email-required'), path: ['confirmEmail'] });
         } else if (!validator.isEmail(val.confirmEmail)) {
-          ctx.addIssue({ code: z.ZodIssueCode.custom, message: t('apply-adult-child:contact-information.error-message.email-valid'), path: ['confirmEmail'] });
+          ctx.addIssue({ code: z.ZodIssueCode.custom, message: t('apply-adult:contact-information.error-message.email-valid'), path: ['confirmEmail'] });
         } else if (val.email !== val.confirmEmail) {
-          ctx.addIssue({ code: z.ZodIssueCode.custom, message: t('apply-adult-child:contact-information.error-message.email-match'), path: ['confirmEmail'] });
+          ctx.addIssue({ code: z.ZodIssueCode.custom, message: t('apply-adult:contact-information.error-message.email-match'), path: ['confirmEmail'] });
         }
       }
 
       if (val.mailingCountry === CANADA_COUNTRY_ID || val.mailingCountry === USA_COUNTRY_ID) {
         if (!val.mailingProvince || validator.isEmpty(val.mailingProvince)) {
-          ctx.addIssue({ code: z.ZodIssueCode.custom, message: t('apply-adult-child:contact-information.error-message.mailing-address.province-required'), path: ['mailingProvince'] });
+          ctx.addIssue({ code: z.ZodIssueCode.custom, message: t('apply-adult:contact-information.error-message.mailing-address.province-required'), path: ['mailingProvince'] });
         }
 
         if (!val.mailingPostalCode || validator.isEmpty(val.mailingPostalCode)) {
-          ctx.addIssue({ code: z.ZodIssueCode.custom, message: t('apply-adult-child:contact-information.error-message.mailing-address.postal-code-required'), path: ['mailingPostalCode'] });
+          ctx.addIssue({ code: z.ZodIssueCode.custom, message: t('apply-adult:contact-information.error-message.mailing-address.postal-code-required'), path: ['mailingPostalCode'] });
         } else if (!isValidPostalCode(val.mailingCountry, val.mailingPostalCode)) {
-          const message = val.mailingCountry === CANADA_COUNTRY_ID ? t('apply-adult-child:contact-information.error-message.mailing-address.postal-code-valid') : t('apply-adult-child:contact-information.error-message.mailing-address.zip-code-valid');
+          const message = val.mailingCountry === CANADA_COUNTRY_ID ? t('apply-adult:contact-information.error-message.mailing-address.postal-code-valid') : t('apply-adult:contact-information.error-message.mailing-address.zip-code-valid');
           ctx.addIssue({ code: z.ZodIssueCode.custom, message, path: ['mailingPostalCode'] });
         }
       }
 
       if (val.copyMailingAddress === false) {
         if (!val.homeAddress || validator.isEmpty(val.homeAddress)) {
-          ctx.addIssue({ code: z.ZodIssueCode.custom, message: t('apply-adult-child:contact-information.error-message.home-address.address-required'), path: ['homeAddress'] });
+          ctx.addIssue({ code: z.ZodIssueCode.custom, message: t('apply-adult:contact-information.error-message.home-address.address-required'), path: ['homeAddress'] });
         }
 
         if (!val.homeCountry || validator.isEmpty(val.homeCountry)) {
-          ctx.addIssue({ code: z.ZodIssueCode.custom, message: t('apply-adult-child:contact-information.error-message.home-address.country-required'), path: ['homeCountry'] });
+          ctx.addIssue({ code: z.ZodIssueCode.custom, message: t('apply-adult:contact-information.error-message.home-address.country-required'), path: ['homeCountry'] });
         }
 
         if (!val.homeCity || validator.isEmpty(val.homeCity)) {
-          ctx.addIssue({ code: z.ZodIssueCode.custom, message: t('apply-adult-child:contact-information.error-message.home-address.city-required'), path: ['homeCity'] });
+          ctx.addIssue({ code: z.ZodIssueCode.custom, message: t('apply-adult:contact-information.error-message.home-address.city-required'), path: ['homeCity'] });
         }
 
         if (val.homeCountry === CANADA_COUNTRY_ID || val.homeCountry === USA_COUNTRY_ID) {
           if (!val.homeProvince || validator.isEmpty(val.homeProvince)) {
-            ctx.addIssue({ code: z.ZodIssueCode.custom, message: t('apply-adult-child:contact-information.error-message.home-address.province-required'), path: ['homeProvince'] });
+            ctx.addIssue({ code: z.ZodIssueCode.custom, message: t('apply-adult:contact-information.error-message.home-address.province-required'), path: ['homeProvince'] });
           }
 
           if (!val.homePostalCode || validator.isEmpty(val.homePostalCode)) {
-            ctx.addIssue({ code: z.ZodIssueCode.custom, message: t('apply-adult-child:contact-information.error-message.home-address.postal-code-required'), path: ['homePostalCode'] });
+            ctx.addIssue({ code: z.ZodIssueCode.custom, message: t('apply-adult:contact-information.error-message.home-address.postal-code-required'), path: ['homePostalCode'] });
           } else if (!isValidPostalCode(val.homeCountry, val.homePostalCode)) {
-            const message = val.homeCountry === CANADA_COUNTRY_ID ? t('apply-adult-child:contact-information.error-message.home-address.postal-code-valid') : t('apply-adult-child:contact-information.error-message.home-address.zip-code-valid');
+            const message = val.homeCountry === CANADA_COUNTRY_ID ? t('apply-adult:contact-information.error-message.home-address.postal-code-valid') : t('apply-adult:contact-information.error-message.home-address.zip-code-valid');
             ctx.addIssue({ code: z.ZodIssueCode.custom, message, path: ['homePostalCode'] });
           }
         }
@@ -227,10 +228,10 @@ export async function action({ context: { session }, params, request }: ActionFu
   }
 
   if (state.editMode) {
-    return redirect(getPathById('$lang/_public/apply/$id/adult-child/review-adult-information', params));
+    return redirect(getPathById('$lang/_public/apply/$id/adult/review-information', params));
   }
 
-  return redirect(getPathById('$lang/_public/apply/$id/adult-child/communication-preference', params));
+  return redirect(getPathById('$lang/_public/apply/$id/adult/communication-preference', params));
 }
 
 export default function ApplyFlowPersonalInformation() {
@@ -353,7 +354,7 @@ export default function ApplyFlowPersonalInformation() {
     })
     .sort((region1, region2) => region1.children.localeCompare(region2.children));
 
-  const dummyOption: InputOptionProps = { children: t('apply-adult-child:contact-information.address-field.select-one'), value: '' };
+  const dummyOption: InputOptionProps = { children: t('apply-adult:contact-information.address-field.select-one'), value: '' };
 
   const postalCodeRequiredContries = [CANADA_COUNTRY_ID, USA_COUNTRY_ID];
   const mailingPostalCodeRequired = selectedMailingCountry !== undefined && postalCodeRequiredContries.includes(selectedMailingCountry);
@@ -373,10 +374,10 @@ export default function ApplyFlowPersonalInformation() {
         <fetcher.Form method="post" noValidate>
           <input type="hidden" name="_csrf" value={csrfToken} />
           <fieldset className="mb-6">
-            <legend className="mb-4 font-lato text-2xl font-bold">{t('apply-adult-child:contact-information.phone-header')}</legend>
-            <p className="mb-4">{t('apply-adult-child:contact-information.add-phone')}</p>
+            <legend className="mb-4 font-lato text-2xl font-bold">{t('apply-adult:contact-information.phone-header')}</legend>
+            <p className="mb-4">{t('apply-adult:contact-information.add-phone')}</p>
             <div className="grid items-end gap-6">
-              <InputField
+              <InputPhoneField
                 id="phone-number"
                 name="phoneNumber"
                 type="tel"
@@ -385,11 +386,11 @@ export default function ApplyFlowPersonalInformation() {
                 autoComplete="tel"
                 defaultValue={defaultState?.phoneNumber ?? ''}
                 errorMessage={errorMessages['phone-number']}
-                label={t('apply-adult-child:contact-information.phone-number')}
+                label={t('apply-adult:contact-information.phone-number')}
                 maxLength={100}
                 aria-describedby="adding-phone"
               />
-              <InputField
+              <InputPhoneField
                 id="phone-number-alt"
                 name="phoneNumberAlt"
                 type="tel"
@@ -398,16 +399,16 @@ export default function ApplyFlowPersonalInformation() {
                 autoComplete="tel"
                 defaultValue={defaultState?.phoneNumberAlt ?? ''}
                 errorMessage={errorMessages['phone-number-alt']}
-                label={t('apply-adult-child:contact-information.phone-number-alt')}
+                label={t('apply-adult:contact-information.phone-number-alt')}
                 maxLength={100}
                 aria-describedby="adding-phone"
               />
             </div>
           </fieldset>
           <fieldset className="mb-6">
-            <legend className="mb-4 font-lato text-2xl font-bold">{t('apply-adult-child:contact-information.email-header')}</legend>
+            <legend className="mb-4 font-lato text-2xl font-bold">{t('apply-adult:contact-information.email-header')}</legend>
             <p id="adding-email" className="mb-4">
-              {t('apply-adult-child:contact-information.add-email')}
+              {t('apply-adult:contact-information.add-email')}
             </p>
             <div className="grid gap-6 md:grid-cols-2">
               <InputField
@@ -419,7 +420,7 @@ export default function ApplyFlowPersonalInformation() {
                 autoComplete="email"
                 defaultValue={defaultState?.email ?? ''}
                 errorMessage={errorMessages['email']}
-                label={t('apply-adult-child:contact-information.email')}
+                label={t('apply-adult:contact-information.email')}
                 maxLength={64}
                 aria-describedby="adding-email"
               />
@@ -432,22 +433,22 @@ export default function ApplyFlowPersonalInformation() {
                 autoComplete="email"
                 defaultValue={defaultState?.email ?? ''}
                 errorMessage={errorMessages['confirm-email']}
-                label={t('apply-adult-child:contact-information.confirm-email')}
+                label={t('apply-adult:contact-information.confirm-email')}
                 maxLength={64}
                 aria-describedby="adding-email"
               />
             </div>
           </fieldset>
           <fieldset className="mb-6">
-            <legend className="mb-4 font-lato text-2xl font-bold">{t('apply-adult-child:contact-information.mailing-address.header')}</legend>
+            <legend className="mb-4 font-lato text-2xl font-bold">{t('apply-adult:contact-information.mailing-address.header')}</legend>
             <div className="space-y-6">
               <InputField
                 id="mailing-address"
                 name="mailingAddress"
                 className="w-full"
-                label={t('apply-adult-child:contact-information.address-field.address')}
+                label={t('apply-adult:contact-information.address-field.address')}
                 maxLength={30}
-                helpMessagePrimary={t('apply-adult-child:contact-information.address-field.address-note')}
+                helpMessagePrimary={t('apply-adult:contact-information.address-field.address-note')}
                 helpMessagePrimaryClassName="text-black"
                 autoComplete="address-line1"
                 defaultValue={defaultState?.mailingAddress ?? ''}
@@ -458,7 +459,7 @@ export default function ApplyFlowPersonalInformation() {
                 id="mailing-apartment"
                 name="mailingApartment"
                 className="w-full"
-                label={t('apply-adult-child:contact-information.address-field.apartment')}
+                label={t('apply-adult:contact-information.address-field.apartment')}
                 maxLength={30}
                 autoComplete="address-line2"
                 defaultValue={defaultState?.mailingApartment ?? ''}
@@ -468,7 +469,7 @@ export default function ApplyFlowPersonalInformation() {
                 id="mailing-country"
                 name="mailingCountry"
                 className="w-full sm:w-1/2"
-                label={t('apply-adult-child:contact-information.address-field.country')}
+                label={t('apply-adult:contact-information.address-field.country')}
                 autoComplete="country"
                 defaultValue={defaultState?.mailingCountry ?? ''}
                 errorMessage={errorMessages['mailing-country']}
@@ -481,7 +482,7 @@ export default function ApplyFlowPersonalInformation() {
                   id="mailing-province"
                   name="mailingProvince"
                   className="w-full sm:w-1/2"
-                  label={t('apply-adult-child:contact-information.address-field.province')}
+                  label={t('apply-adult:contact-information.address-field.province')}
                   defaultValue={defaultState?.mailingProvince ?? ''}
                   errorMessage={errorMessages['mailing-province']}
                   options={[dummyOption, ...mailingRegions]}
@@ -493,7 +494,7 @@ export default function ApplyFlowPersonalInformation() {
                   id="mailing-city"
                   name="mailingCity"
                   className="w-full"
-                  label={t('apply-adult-child:contact-information.address-field.city')}
+                  label={t('apply-adult:contact-information.address-field.city')}
                   maxLength={100}
                   autoComplete="address-level2"
                   defaultValue={defaultState?.mailingCity ?? ''}
@@ -504,7 +505,7 @@ export default function ApplyFlowPersonalInformation() {
                   id="mailing-postal-code"
                   name="mailingPostalCode"
                   className="w-full"
-                  label={mailingPostalCodeRequired ? t('apply-adult-child:contact-information.address-field.postal-code') : t('apply-adult-child:contact-information.address-field.postal-code-optional')}
+                  label={mailingPostalCodeRequired ? t('apply-adult:contact-information.address-field.postal-code') : t('apply-adult:contact-information.address-field.postal-code-optional')}
                   maxLength={100}
                   autoComplete="postal-code"
                   defaultValue={defaultState?.mailingPostalCode}
@@ -515,10 +516,10 @@ export default function ApplyFlowPersonalInformation() {
             </div>
           </fieldset>
           <fieldset className="mb-8">
-            <legend className="mb-4 font-lato text-2xl font-bold">{t('apply-adult-child:contact-information.home-address.header')}</legend>
+            <legend className="mb-4 font-lato text-2xl font-bold">{t('apply-adult:contact-information.home-address.header')}</legend>
             <div className="space-y-6">
               <InputCheckbox id="copyMailingAddress" name="copyMailingAddress" value="copy" checked={copyAddressChecked} onChange={checkHandler}>
-                {t('apply-adult-child:contact-information.home-address.use-mailing-address')}
+                {t('apply-adult:contact-information.home-address.use-mailing-address')}
               </InputCheckbox>
               {!copyAddressChecked && (
                 <>
@@ -526,8 +527,8 @@ export default function ApplyFlowPersonalInformation() {
                     id="home-address"
                     name="homeAddress"
                     className="w-full"
-                    label={t('apply-adult-child:contact-information.address-field.address')}
-                    helpMessagePrimary={t('apply-adult-child:contact-information.address-field.address-note')}
+                    label={t('apply-adult:contact-information.address-field.address')}
+                    helpMessagePrimary={t('apply-adult:contact-information.address-field.address-note')}
                     helpMessagePrimaryClassName="text-black"
                     maxLength={30}
                     autoComplete="address-line1"
@@ -539,7 +540,7 @@ export default function ApplyFlowPersonalInformation() {
                     id="home-apartment"
                     name="homeApartment"
                     className="w-full"
-                    label={t('apply-adult-child:contact-information.address-field.apartment')}
+                    label={t('apply-adult:contact-information.address-field.apartment')}
                     maxLength={30}
                     autoComplete="address-line2"
                     defaultValue={defaultState?.homeApartment ?? ''}
@@ -549,7 +550,7 @@ export default function ApplyFlowPersonalInformation() {
                     id="home-country"
                     name="homeCountry"
                     className="w-full sm:w-1/2"
-                    label={t('apply-adult-child:contact-information.address-field.country')}
+                    label={t('apply-adult:contact-information.address-field.country')}
                     autoComplete="country"
                     defaultValue={defaultState?.homeCountry ?? ''}
                     errorMessage={errorMessages['home-country']}
@@ -562,7 +563,7 @@ export default function ApplyFlowPersonalInformation() {
                       id="home-province"
                       name="homeProvince"
                       className="w-full sm:w-1/2"
-                      label={t('apply-adult-child:contact-information.address-field.province')}
+                      label={t('apply-adult:contact-information.address-field.province')}
                       defaultValue={defaultState?.homeProvince ?? ''}
                       errorMessage={errorMessages['home-province']}
                       options={[dummyOption, ...homeRegions]}
@@ -574,7 +575,7 @@ export default function ApplyFlowPersonalInformation() {
                       id="home-city"
                       name="homeCity"
                       className="w-full"
-                      label={t('apply-adult-child:contact-information.address-field.city')}
+                      label={t('apply-adult:contact-information.address-field.city')}
                       maxLength={100}
                       autoComplete="address-level2"
                       defaultValue={defaultState?.homeCity ?? ''}
@@ -585,7 +586,7 @@ export default function ApplyFlowPersonalInformation() {
                       id="home-postal-code"
                       name="homePostalCode"
                       className="w-full"
-                      label={homePostalCodeRequired ? t('apply-adult-child:contact-information.address-field.postal-code') : t('apply-adult-child:contact-information.address-field.postal-code-optional')}
+                      label={homePostalCodeRequired ? t('apply-adult:contact-information.address-field.postal-code') : t('apply-adult:contact-information.address-field.postal-code-optional')}
                       maxLength={100}
                       autoComplete="postal-code"
                       defaultValue={defaultState?.homePostalCode ?? ''}
@@ -599,34 +600,28 @@ export default function ApplyFlowPersonalInformation() {
           </fieldset>
           {editMode ? (
             <div className="flex flex-wrap items-center gap-3">
-              <Button variant="primary" id="continue-button" disabled={isSubmitting} data-gc-analytics-customclick="ESDC-EDSC:CDCP Online Application Form-Adult_Child:Save - Contact information click">
-                {t('apply-adult-child:contact-information.save-btn')}
+              <Button variant="primary" id="continue-button" disabled={isSubmitting} data-gc-analytics-customclick="ESDC-EDSC:CDCP Online Application Form-Adult:Save - Contact information click">
+                {t('apply-adult:contact-information.save-btn')}
               </Button>
-              <ButtonLink
-                id="back-button"
-                routeId="$lang/_public/apply/$id/adult-child/review-adult-information"
-                params={params}
-                disabled={isSubmitting}
-                data-gc-analytics-customclick="ESDC-EDSC:CDCP Online Application Form-Adult_Child:Cancel - Contact information click"
-              >
-                {t('apply-adult-child:contact-information.cancel-btn')}
+              <ButtonLink id="back-button" routeId="$lang/_public/apply/$id/adult/review-information" params={params} disabled={isSubmitting} data-gc-analytics-customclick="ESDC-EDSC:CDCP Online Application Form-Adult:Cancel - Contact information click">
+                {t('apply-adult:contact-information.cancel-btn')}
               </ButtonLink>
             </div>
           ) : (
             <div className="flex flex-row-reverse flex-wrap items-center justify-end gap-3">
-              <Button variant="primary" id="continue-button" disabled={isSubmitting} data-gc-analytics-customclick="ESDC-EDSC:CDCP Online Application Form-Adult_Child:Continue - Contact information click">
-                {t('apply-adult-child:contact-information.continue')}
+              <Button variant="primary" id="continue-button" disabled={isSubmitting} data-gc-analytics-customclick="ESDC-EDSC:CDCP Online Application Form-Adult:Continue - Contact information click">
+                {t('apply-adult:contact-information.continue')}
                 <FontAwesomeIcon icon={isSubmitting ? faSpinner : faChevronRight} className={cn('ms-3 block size-4', isSubmitting && 'animate-spin')} />
               </Button>
               <ButtonLink
                 id="back-button"
-                routeId={[MARITAL_STATUS_CODE_COMMONLAW, MARITAL_STATUS_CODE_MARRIED].includes(Number(maritalStatus)) ? '$lang/_public/apply/$id/adult-child/partner-information' : '$lang/_public/apply/$id/adult-child/applicant-information'}
+                routeId={[MARITAL_STATUS_CODE_COMMONLAW, MARITAL_STATUS_CODE_MARRIED].includes(Number(maritalStatus)) ? '$lang/_public/apply/$id/adult/partner-information' : '$lang/_public/apply/$id/adult/applicant-information'}
                 params={params}
                 disabled={isSubmitting}
-                data-gc-analytics-customclick="ESDC-EDSC:CDCP Online Application Form-Adult_Child:Back - Contact information click"
+                data-gc-analytics-customclick="ESDC-EDSC:CDCP Online Application Form-Adult:Back - Contact information click"
               >
                 <FontAwesomeIcon icon={faChevronLeft} className="me-3 block size-4" />
-                {t('apply-adult-child:contact-information.back')}
+                {t('apply-adult:contact-information.back')}
               </ButtonLink>
             </div>
           )}
