+++ frontend/app/routes/$lang/_public/apply/$id/child/applicant-information.tsx
@@ -223,7 +223,7 @@ export async function action({ context: { session }, params, request }: ActionFu
     return redirect(getPathById('$lang/_public/apply/$id/child/partner-information', params));
   }
 
-  return redirect(getPathById('$lang/_public/apply/$id/child/personal-information', params));
+  return redirect(getPathById('$lang/_public/apply/$id/child/contact-information', params));
 }
 
 export default function ApplyFlowApplicationInformation() {
