+++ frontend/app/routes/$lang/_public/apply/$id/child/contact-information.tsx
@@ -71,7 +71,7 @@ export async function loader({ context: { session }, params, request }: LoaderFu
 }
 
 export async function action({ context: { session }, params, request }: ActionFunctionArgs) {
-  const log = getLogger('apply/personal-information');
+  const log = getLogger('apply/contact-information');
 
   const state = loadApplyChildState({ params, request, session });
   const t = await getFixedT(request, handle.i18nNamespaces);
